src/utils/generate/fresh_alert_api/pyproject.toml
[tool.poetry]
name = "fresh-alert-api-client"
version = "1.2.1"
description = "A client library for accessing FreshAlert API"
authors = []
readme = "README.md"
packages = [
    { include = "fresh_alert_api_client" },
]
include = ["CHANGELOG.md", "fresh_alert_api_client/py.typed"]

[tool.poetry.dependencies]
python = "^3.9"
httpx = ">=0.23.0,<0.29.0"
attrs = ">=22.2.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
